requirements.txt
Mastodon.py
python-dotenv
kafka-python
pyspark==3.5.1
pymongo
cassandra-driver